[backend] add a retry event for failed gbininfo requests

+++ src/backend/BSSched/EventHandler.pm
@@ -295,7 +295,14 @@ sub event_scanprjbinaries {
   my $remoteprojs = $gctx->{'remoteprojs'};
   if (defined($projid) && defined($repoid)) {
     my $prp = "$projid/$repoid";
-    delete $gctx->{'remotegbininfos'}->{"$prp/$myarch"};
+    my $arch = $ev->{'arch'} || $myarch;
+    delete $gctx->{'remotegbininfos'}->{"$prp/$arch"};
+    if ($ev->{'arch'}) {
+      # remote gbininfo retry event
+      my $changed_med = $gctx->{'changed_med'};
+      $changed_med->{$prp} = 2;
+      return;
+    }
     return if $remoteprojs->{$projid};
     if (defined($packid)) {
       unlink("$prp/$myarch/$packid/.bininfo");
